Added basic ui for person list card

+++ app/src/main/java/basaraba/adndrii/movieguide/features/main/persons/PersonsUiEvent.kt
@@ -3,5 +3,5 @@ package basaraba.adndrii.movieguide.features.main.persons
 
 sealed interface PersonsUiEvent {
     object ReloadPersonsScreen : PersonsUiEvent
-    data class ShowPersonDetails(val id: Int) : PersonsUiEvent
+    data class ShowPersonDetails(val id: Long) : PersonsUiEvent
 }
